vanilla_planner_poc/requirements.txt
azure-ai-agentserver-core==1.0.0b16
azure-ai-projects
azure-identity
openai
python-dotenv
gremlinpython>=3.7.0

# Azure Monitor / OpenTelemetry
azure-monitor-opentelemetry-exporter>=1.0.0b46
opentelemetry-sdk>=1.39.0
opentelemetry-api>=1.39.0
